Absorb another branch to use can_cast_safely
